ingestion/requirements.txt
eumdac>=2.0.1
netCDF4>=1.6.5
numpy>=1.26.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
